Move prefix_pool to sdk
This removes the last of the minor direct dependencies between
sdk and controlplane (but not controlplane/api).
